SVM_SGD_NN notebook updated with additional NN w/ SGD
No material changes other than addition of NN with SGD optimizer. Was originally going to ensemble NNs since SGD wouldn't run on local machine, but tuned the model and got it to run. Added this model to compare to previous best model (1 model before it). SGD wins in terms of generalizability with accuracy as success metric.